converted all api routes
