+++ app/api/user/postAnswer/route.js
@@ -1,6 +1,6 @@
-import { ifDocExists, connectDB } from "../../../middleware";
-import { StudentAnswers, StudentSchema } from "../../../models/users/User";
-ifDocExists
+import { ifDocExists, connectDB } from "../../../../middleware";
+import { StudentAnswers, StudentSchema } from "../../../../models/users/User";
+import { NextResponse } from "next/server";
 
 const handler = async (req, res) => {
   if (req.method === 'POST') {
@@ -17,12 +17,13 @@ const handler = async (req, res) => {
       lesson: lesson
     })
     const result = ifDocExists(user, 'answers', quizAnswer, StudentSchema, res)
-    return result
+    return NextResponse.json(result, { status: 200 })
 
 
   } else {
-    res.status(422).send('method not supported')
+    return NextResponse.json('method not supported', { status: 422 })
   }
 }
 
-export default connectDB(handler)
+const wrappedHandler = connectDB(handler)
+export { wrappedHandler as POST }
