+++ app/api/user/postVocab/route.js
@@ -1,5 +1,6 @@
-import { ifDocExists, connectDB } from "../../../middleware";
-import { StudentSchema, StudentVocab } from "../../../models/users/User";
+import { ifDocExists, connectDB } from "../../../../middleware";
+import { StudentSchema, StudentVocab } from "../../../../models/users/User";
+import { NextResponse } from "next/server";
 
 const handler = async (req, res) => {
     if (req.method === 'POST') {
@@ -12,11 +13,13 @@ const handler = async (req, res) => {
             term: term
         })
         const result = ifDocExists(user, 'vocab', vocabUpdate, StudentSchema, res)
-        return result
+        return NextResponse.json(result, { status: 200 })
 
     } else {
-        res.status(422).send('method not supported')
+        return NextResponse.json('method not supported', { status: 422 })
     }
 }
 
-export default connectDB(handler)
+const wrappedHandler = connectDB(handler)
+
+export { wrappedHandler as POST }
