+++ app/not-found.js
@@ -1,18 +1,17 @@
 'use client'
-import { MainContainer } from '../components/meta'
 import { BlueButton } from '../components/atomic'
+import { Title, Intro } from '../components/layout'
 
 export default function NotFound() {
     return (
-        <MainContainer
-            titleText='404'
-            introText="This page doesn't exist... yet"
-        >
+        <>
+            <Title>404</Title>
+            <Intro>This page doesn't exist... yet</Intro>
             <p>Do you want it to?</p>
             <BlueButton
                 id='404'
                 onClick={() => alert('duly noted')}
             >Yes</BlueButton>
-        </MainContainer>
+        </>
     )
 }
