--- middleware/fetchers/fetchLessonPage.js
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-export default async function fetchLessonPage(sphere, course, lesson) {
-    try {
-        const response = await axios.get(`${process.env.BASE_URL}/api/lesson/lessonPage`, { params: { sphere: sphere, course: course, lesson: lesson } });
-        const fileContents = response.data.content;
-        return fileContents
-    } catch (error) {
-        console.error('Error fetching lesson page:');
-    }
-}
