+++ package.json
@@ -27,7 +27,7 @@
     "mongodb": "^5.6.0",
     "mongodb-client-encryption": "^2.9.0",
     "mongoose": "^7.3.1",
-    "next": "^13.4.12",
+    "next": "13.4.13-canary.4",
     "next-auth": "^4.22.1",
     "next-compose-plugins": "^2.2.1",
     "next-mdx-remote": "^4.4.1",
