--- pages/api/admin/env.js
@@ -1,5 +1,0 @@
-export default function handler(req, res) {
-    const { BASE_URL } = process.env;
-
-    res.status(200).json({ url: BASE_URL });
-}
