--- pages/api/authold/[...nextauth].js
@@ -1,27 +1,0 @@
-import NextAuth from "next-auth"
-import GoogleProvider from "next-auth/providers/google"
-import EmailProvider from "next-auth/providers/email"
-import { MongoDBAdapter } from "@auth/mongodb-adapter"
-import clientPromise from "./lib/mongodb"
-
-
-export const authOptions = {
-    // Configure one or more authentication providers
-    providers: [
-
-        GoogleProvider({
-            clientId: process.env.GOOGLE_ID,
-            clientSecret: process.env.GOOGLE_SECRET,
-        }),
-
-        EmailProvider({
-            server: process.env.EMAIL_SERVER,
-            from: process.env.EMAIL_FROM
-        }),
-        // ...add more providers here
-    ],
-
-    adapter: MongoDBAdapter(clientPromise, { collections: 'school', databaseName: 'school' })
-}
-
-export default NextAuth(authOptions)
